Move parser logic from parseEvent to parser
