+++ src/parseEvent.js
@@ -1,40 +1,5 @@
-import {
-    parseGAMRequest,
-    parseYouTubeRequest,
-} from './parser'
+import { parseRequest } from './parser'
 
-function cleanRequest(request) {
-    return request
-        // Remove any leading / trailing whitespace
-        .trim()
-        // Remove leading string that may be included when copying from browser
-        .replace("Request URL: ", "");
-}
-
-function isYouTubeRequest(request) {
-    if (request.startsWith("https://www.youtube.com/embed/")) {
-        return true;
-    }
-}
-
-function isGAMRequest(request) {
-    if (request.startsWith("https://securepubads.g.doubleclick.net/gampad/ads")) {
-        return true;
-    }
-}
-
-function parseRequest(request, truncateValues, ignoreValues) {
-    if (request && request.length > 0) {
-        if (isYouTubeRequest(request)) {
-            return parseYouTubeRequest(request, truncateValues, ignoreValues)
-        } else if (isGAMRequest(request)) {
-            return parseGAMRequest(request, truncateValues, ignoreValues);  
-        } else {
-            return "";
-        }
-    }
-    return "";
-}
 
 function parseEvent(event) {
     event.preventDefault();
@@ -45,8 +10,8 @@ function parseEvent(event) {
     const ignoreValues = !includeValues;
 
     return [
-        parseRequest(cleanRequest(request1), truncateValues, ignoreValues),
-        parseRequest(cleanRequest(request2), truncateValues, ignoreValues),
+        parseRequest(request1, truncateValues, ignoreValues),
+        parseRequest(request2, truncateValues, ignoreValues),
     ];
 }
